+++ src/algorithms/enumeration/index.ts
@@ -1,2 +1,3 @@
 export * from './BreadthFirstEnum';
 export * from './DepthFirstEnum';
+export * from './NeighborGenerator';