+++ src/algorithms/optimization/index.ts
@@ -1,2 +1,3 @@
 export * from './ComplexityAnalyzer';
 export * from './PerformanceOptimizer';
+export * from './ComplexityEstimate';