+++ CHANGELOG.md
@@ -8,3 +8,12 @@
  - *Dev only features* : Migrate eslint, refactor imports, CI for developers.
 
 <hr>
+
+### 0.1.0 (RemyK)
+ - Chore : update project metadata and documentation *(11/07/2025)*
+ - Feat : added 2 new files : `ComplexityEstimate` and `NeighborGenerator` *(12/07/2025)*
+ - Feat : code completed for `BreadthFirstEnum`, `DepthFirstEnum`, `ComplexityAnalyzer` and `PerformanceOptimizer`
+ - Refactor : import update
+ - Chore : `package.json` update
+
+<hr>